avoid run get_logger
